+++ fortuneboot-service/src/main/java/com/fortuneboot/service/fortune/FortunePayeeService.java
@@ -1,16 +1,18 @@
 package com.fortuneboot.service.fortune;
 
 import com.fortuneboot.common.enums.fortune.BillTypeEnum;
+import com.fortuneboot.common.exception.ApiException;
+import com.fortuneboot.common.exception.error.ErrorCode;
 import com.fortuneboot.domain.command.fortune.FortunePayeeAddCommand;
 import com.fortuneboot.domain.command.fortune.FortunePayeeModifyCommand;
 import com.fortuneboot.domain.entity.fortune.FortunePayeeEntity;
 import com.fortuneboot.domain.query.fortune.FortunePayeeQuery;
 import com.fortuneboot.factory.fortune.FortunePayeeFactory;
 import com.fortuneboot.factory.fortune.model.FortunePayeeModel;
+import com.fortuneboot.repository.fortune.FortuneBillRepository;
 import com.fortuneboot.repository.fortune.FortunePayeeRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
@@ -28,6 +30,8 @@ public class FortunePayeeService {
     private final FortunePayeeRepository fortunePayeeRepository;
 
     private final FortunePayeeFactory fortunePayeeFactory;
+
+    private final FortuneBillRepository fortuneBillRepository;
 
     public List<FortunePayeeEntity> getList(FortunePayeeQuery query) {
         return fortunePayeeRepository.list(query.addQueryCondition());
@@ -68,6 +72,10 @@ public class FortunePayeeService {
     }
 
     public void remove(Long bookId, Long payeeId) {
+        Boolean used = fortuneBillRepository.existByPayeeId(payeeId);
+        if (used){
+            throw new ApiException(ErrorCode.Business.PAYEE_ALREADY_USED);
+        }
         FortunePayeeModel fortunePayeeModel = fortunePayeeFactory.loadById(payeeId);
         fortunePayeeModel.checkBookId(bookId);
         fortunePayeeModel.deleteById();
